+++ lib/src/inputs/phone/phone_number_model.dart
@@ -1,0 +1,9 @@
+class PhoneNumber {
+  const PhoneNumber({
+    this.dialCode,
+    this.number,
+  });
+
+  final String? dialCode;
+  final String? number;
+}